requirements.txt
# Core Framework
fastapi
uvicorn

# Data Validation
pydantic
pydantic[email]

# AI/ML
google-generativeai
langchain
langchain-google-genai
chromadb
langchain-core
pinecone-client

# Web Scraping
beautifulsoup4
requests
lxml
requests-html
nest-asyncio
duckduckgo-search

# Utilities
python-dotenv
slowapi
tenacity
crawl4ai
tqdm
sentence-transformers
openai
